Fix band mapping in radconstants and albedo
Band mapping in the shortwave changed between RRTMG and RRTMGP, but the
albedo mapping was copied directly over from the RRTMG implementation.
This caused some big differences in the reflected shortwave. This is
now fixed, and the band mapping is handled a little more intelligently
in the set_albedo routine now.