Cosmetic and functional fixes to CRM interface
It seemed that crm_physics was called at an odd place within the if block for l_tracer_aero in tphysbc() of physpkg.F90, so I moved it out and cleaned up the indentation and comments to make the code structure clearer. This sent me into a whirlwind of other adjustments in crm_physics.F90 and crm_module.F90. I added the factor_xyt variable so that this factor wasn't constantly being recalculated multiple times for each crm instance. I also added a lot of comments.

One functional change that I have not tested yet is an additional check for "use_ECPP" at line 2760 in physpkg.F90. I think this is necessary since we want the 2-moment configuration to work with AND without ECPP.

I also moved around the calls for the SP_CRM_BULK and SP_PHYS_BYPASS options, but this does not change the default behavior. These options are still a work in progress.

	modified:   components/cam/src/physics/cam/crm_physics.F90
	modified:   components/cam/src/physics/cam/physpkg.F90
	modified:   components/cam/src/physics/crm/crm_module.F90