Bug fix for CRM setperturb()
Awhile back I identified a problem with how the CRM perturbations were being set. Specifically, the intrinsic random number generator method is VERY compiler dependent, and it turned out that the magnitude of perturbations was effectively zero. There was a related problem in that regression tests weren't passing because the perturbations were not reproducible. So I implemented a new RNG using the "Mersenne Twister" algorithm. This solved the reproducibility issue so our regression tests were solved, but I now realize that I introduced an even worse problem by inadvertanty bypassing the call to setperturb() altogether! The reason for this was that I naively used the "is_first_step()" function, thinking that it would be true during the first CRM integration. However, I failed to recognize that the first CRM call happens when nstep=1, whereas is_first_step() only returns true when nstep=0. Realizing this I changed crm() to check that the GCM timestep is <=1. This ensured that setperturb() was being called, but in single column model tests I noticed there was still an issue where convective overturning was not occurring within the CRM grid. To remedy this I modified setperturb() to use a deeper level and apply larger perturbations. instead of a hard-coded level number, the subroutine finds all levels under a specified pressure level (currently 700 hPa).

	modified:   components/cam/src/physics/cam/crm_physics.F90
	modified:   components/cam/src/physics/crm/crm_module.F90
	modified:   components/cam/src/physics/crm/setperturb.F90