Added new SP "AQUA" and "AQUAP" compsets
The AQUAP compset is meant to be similar to the old CAM aquaplanet compset, which currently has some bugs in that the topo file is not corect. The other "AQUA" compset is more like a typical CAM F compset, where the input SST, topo, and domain files can be specified manually. Note the only difference is in how the OCN component is specified (%DOM vs %AQP1).

        modified:   components/cam/cime_config/config_compsets.xml