Isolate CAMRT source from main CAM source
The old build behavior included the CAMRT radiative transfer code into
the main CAM source, and configure would overwrite the source paths to
corresponding files in the RRTMG source directory when RRTMG was
selected as the radiative transfer code to build with the model. This
was somewhat confusing, so in order to improve transparency and to
isolate separate packages, the CAMRT-specific source files (especially
those with counterparts in RRTMG) have been moved to a separate
directory that should entirely contain the interface to the (outdated)
CAMRT radiative transfer package. At some point, support for this
package may be dropped from the model entirely, and this should make
that easier to do in the future. The configure script for CAM has also
been edited to allow for this change, and to prepare for the addition of
the RRTMGP code in future commits.