Move CRM micro_fields handling into crm_module
Move all handling of the CRM micro_fields array (used by the CRM
microphysics routines) down into the crm() subroutine in crm_module.
Since all this is specific to the CRM, there is really no reason for it
to exist at the crm_physics level. Also, by pushing this down and
instead explicitly adding microphysics fields to the crm_state as
pointers, we can point the physics buffer references to these directly
to the fields in crm_state, and avoid a little bit of extra data
copying.