Added SP_RAD_NO_AVG option
The SP_RAD_NO_AVG option disables the time averaging of CRM fields for the radiative calculation. Some analysis of weak surface fluxes suggested that there was insufficient net shortwave radiation at the surface. We realized that the time averaging of clouds in the CRM is partly to blame since it smears the clouds across the domain and reduces the occurrence of clear skies that allow direct warming of the surface by solar radiation.

	modified:   components/cam/bld/configure
	modified:   components/cam/src/physics/crm/crm_module.F90