Fix crm presi initialization bug; minor fixes for titan debug build
The crm interface pressure array (presi) was not properly initialized
because it was not used prior to the recent update to tke_full.F90.
This has now been fixed by adding "pint" as an input to the crm
subroutine and mirroring the calculation of pres.

Additional changes were required for the model to compile on titan
with debug flags:
     * add #ifdef m2005 block in crm_physics.F90
     * delete dummy Get_reff[ci] functions in
       .../src/physics/crm/MICRO_SAM1MOM/microphysics