Major SP Bug fixes
Two major bugs were identified that were affecting regression tests and the fidelity of the model climate

The first was how the crm_angle was set. I mistakenly had this variable set in an if state for is_first_step(), which meant that it was not set on restarts, making the model non-BFB! This has been corrected so the crm_angle variable is always set.

The second bug fix relates to aerosol scavenging. While working on an experimental bypass of non-CRM phyics I naively zeroed out the rain productino variables in the pbuf without realizing that these are need for aerosol scavenging that is calculated after the CRM runs (only for prognostic aerosols).

	modified:   components/cam/src/physics/cam/crm_physics.F90