Changed how tend variable is treated
I realized the "tend" variable is no longer needed by crm_physics_tend() so I removed it. I also changed how tend_save was handled. it is still allocated and deallocated, but I changed it to a direct copy instead of going through each variable separately. The only difference I can tell rom the old way was that the flx_net variable was not being saved and remembered correctly. However, it doesnt seem that this variable is used anywhere and may be a relic of the old eulerian spectral core. I also removed the if statement check for SPCAM in the crm_save and crm_remember routines since this is redundant.