Large update to crm_bulk.F90
The update focuses on crm_bulk_aero_mix_nuc(). I'm starting to understand the aerosol routines a bit better, and I realized that having aerosol activiation happen inside the call dropmixnuc() is better than calling activate_modal() directly. So for now, crm_bulk_aerosol_activation() is commented out, and I'm focusing on crm_bulk_aero_mix_nuc(). I still need to make sure various quantities are directly derived from crm fields that are stored in the pbuf, so there is lots more work to do. I've tried to remove stuff that minghaui added which is not being used, but there's a lot of it that I don;'t fully understand, so I've also left a lot of relics in.