Remove qtot from crm call
Remove qtot field from the crm call. This field is not being used
anymore, and does not need to be passed in and out of the crm. Made this
a local variable in crm_module to allow the lines containing the
associated calculation to remain in case we want to revisit this.