Fix initialization in crm_input_module
The previous commit contained an error in the initialization section of
the module, where member fields were not referred to as members of the
class but rather as standalone variables themselves. This prevented the
model from building, because these variables were not declared. This
commit fixes the initialization (i.e., this%zmid = 0 to replace zmid =
0).