Allow setting cld in radiation from CRM
Allow setting the cloud fraction variable "cld" in radiation_tend from
the saved crm_rad%cld field, calculated in the CRM by averaging the
binary occurence of cloud in time and possible in space (i.e., reduced
rad option). This should force mcica sampling to be done on the CRM
optical properties, rather than assuming that clouds are fully resolved.
We may need to adjust the thresholds for considering what is "cloud" in
order to do this.