Make qtot threshold NOT a magic number
Move the threshold for considering qtot high enough to call a column
"cloudy" to a parameter rather than a magic number. NOTE: this threshold
maybe needs to change or go away when using crm_rad%cld for the cloud
fraction? This was previously used to get the CRM cloud fraction for rad
calculations (either 1 or 0).