Fix radiation timestep in RRTMG
Radiation needs to be done every timestep for SP runs. Previously, this
was set in radiation_tend() by overwriting dosw and dolw, which are set
in radiation_do() by checking the timestep against iradsw and iradlw.
However, the coszrs calulation is based on iradsw, and I believe the
land model uses iradsw to determine how often to update surface albedo.
So, there is a possibility of a mismatch in surface albedo and radiation
with this approach, and there is a baked-in mismatch in coszrs (because
if iradsw > 1, the coszrs calculation is based on a longer time period
than it should be, so radiation sees a coszrs that looks like it is
earlier in the day). This fixes these issues by setting iradsw and
iradlw in radiation_init() to be 1 if doing SP.