Add logic to ensure ntavg2_ss is multiple of ntavg1_ss (fixes #47)
In ECPP, the level-2 averaging period (ntavg2_ss) must be a multiple
of the level-1 averaging period (ntavg1_ss). By default ntavg1_ss
is 10 minutes or dtime, whichever is shorter. This breaks when dtime
is not a multiple of 10 minutes. This fix adjusts ntavg1_ss to be the
smallest time larger than 10 minutes that evenly divides ntavg2_ss when
dtime is longer than 10 minutes.

[BFB] for previously working configurations.