Add check for all night columns
Add a check to see if chunk contains only nighttime columns, and if so
have the shortwave radiation driver zero out fluxes and heating rates
and immediately return without calling any RRTMGP routines. This is
needed because RRTMGP needs to be run on only daytime columns (fails
if any night columns are provided), and if we have all night columns
then nday = 0, so allocation routines fail because we try to allocate
size zero array objects. RRTMG seems to get around this by setting
coszrs to a small number when it is zero, but I am not sure how the
daytime reduction of columns succeeds for nday = 0 in that case. Maybe
because the RRTMG interface does not use any dynamic memory allocation.