Add check to clip temperatures going into RRTMGP
Clip temperatures going into RRTMGP if they are outside the valid range of the
RRTMGP absorption coefficient data provided in the input files. This is to
handle cases where temperatures may be just outside the valid range, such as
for the RCEMIP case. Without clipping values, the code will fail with the error
message that the temperatures are outside the valid range. A better thing to do
than clipping here might be to obtain absorption coefficient data that extends
to a larger temperature range, but this allows us to run the RCEMIP experiment
for now.