Re-implement mean-state acceleration for superparameterization
Reimplements mean-state acceleration for SP-E3SM. This was originally
implemented in Jan 2018, but found to be unstable. The mode of
instability was similar to other modes of instability in the
standard, unaccelerated SP-E3SM (specifically, negative thermodynamic
variables in the CRM) which have since been tamed. Thus, this
warrants additional consideration. Initial (re) commit - expect
bugs. This also includes unit-tests that were used for offline testing.