More tweaks, but no solution
I've been tweaking the phys_hyperviscosity methods a lot, but I can't get anything that runs properly. I thought had a "working" version, but the grid imprinting actually got worse! After isolating the packing/unpacking part of the code it appears that the packing and unpacking works differently for the tendencies of temperature and water tracers. Simply packing and unpacking the water vapor tendency without any hyperviscosity leads to larger grid imprinting. I think this might be a subtle weihting issue, but I'm running out of ideas. I tried several ways to apply pressure weighting, as is done in the dynamics, but that only caused unexplained seg faults. I also tried replacing the dummy "phys_elem" with an "elem" variable passed down from cam_comp.F90, but this didnt really change anything. I've left this option in place for now, just in case I want to try it again.

The SP_DUMMY_HYPERVIS option allows a test of what I'm talking about. It takes the qv tendency, packs the edge buffer, communicates the boundaries, and unpacks the edge buffer. This includes the mass weighting that is needed.

	modified:   components/cam/src/control/cam_comp.F90
	modified:   components/cam/src/physics/cam/phys_hyperviscosity_mod.F90
	modified:   components/cam/src/physics/cam/physpkg.F90
	modified:   components/cam/src/physics/cam/tphysbc_sp_mod.F90
	modified:   components/homme/src/share/element_mod.F90