First attempt of implementing 1x1 phys grid
The idea is to have 1 physics column per dynamics element. For simplicity, the current aim is to have the physics state correspond to the average of the middle GLL nodes and apply the physics tendencies in a piece-linear fashion back to all nodes of the element after the physics. Ultimately we'll want to use the whole element. I've already run into some issues with the initialization of components in the coupler. Remaking the domain files to match the physics grid seems to have fixed this problem, but further mapping files will be needed to use the ocean/ice models.

Note that this version does not run.

	modified:   cime/config/e3sm/allactive/config_pesall.xml
	modified:   cime/config/e3sm/config_grids.xml
	modified:   cime/src/drivers/mct/main/seq_domain_mct.F90
	modified:   components/cam/bld/config_files/horiz_grid.xml
	modified:   components/cam/bld/namelist_files/namelist_defaults_cam.xml
	modified:   components/cam/src/control/ncdio_atm.F90
	modified:   components/cam/src/cpl/atm_comp_mct.F90
	modified:   components/cam/src/dynamics/se/dp_coupling.F90
	modified:   components/cam/src/dynamics/se/dyn_grid.F90
	modified:   components/cam/src/dynamics/se/stepon.F90
	modified:   components/cam/src/physics/cam/phys_grid.F90
	modified:   components/cam/src/physics/rrtmgp/radiation.F90
	modified:   components/homme/test/template/makegrid.job