Fixes for running pg1 with less than num_dyn tasks
The initial prototype was required to use NTASK equal to the number of physics columns. The complication was related to the MPI calls in get_horiz_grid_e(), and specifically to the displacement argument required for mpi_allgatherv(). Initially I set the displacement equal to the global element ID, which conveniently sorted the columns into an order that worked. However, when I tried changing the number of tasks to num_dyn/2 I hit a wall because the elements appear to be randomly assigned to the MPI tasks. The first part of the solution was to reformulate the displacement variable to reflect the (possibly varying) number of elements assigned to each task. This correctly gathered all the data, but the data was not ordered according to the global element id, so I had to add a sorting step.

	modified:   components/cam/src/dynamics/se/dyn_grid.F90