Added gll_grid.F90
These changes are my initial attempts to add output on the dynamics grid (i.e. gll) when the "pg1" grid is used. I think there is probably a better way to do all this. The current approach is to mimic the normal set up of the phys grid with different names, which then can be used by the addfld() and outfld() calls at the end of stepon_run2() in order to get data on the GLL grid in the history file.

note - this feature is currently not working due to run time failures.

 Changes to be committed:
	modified:   components/cam/src/control/cam_comp.F90
	modified:   components/cam/src/dynamics/se/inital.F90
	modified:   components/cam/src/dynamics/se/stepon.F90
	new file:   components/cam/src/physics/cam/gll_grid.F90
	modified:   components/cam/src/physics/cam/phys_grid.F90