Working version of pg1 w/ dyn grid output
the output variables that start with "dyn_" are now able to be output in the same file as phys grid variables.

Although I used "dyn" in the variable names, I used "gll" to describe the grid. For instance, the lat variable that corresponds to the dynamics grid is "gll_lat". I should probably make this all consistent, but I keep waffling on what to name all these things.

The test runs with global RCE have been stable and consistent with np4 runs. Initial comparisons of dyn and phys grid outputs show that the global means of the two states are evolving in step with each other, and not drifting apart.

	modified:   components/cam/src/control/cam_comp.F90
	modified:   components/cam/src/dynamics/se/dp_coupling.F90
	modified:   components/cam/src/dynamics/se/dyn_grid.F90
	modified:   components/cam/src/dynamics/se/stepon.F90
	modified:   components/cam/src/physics/cam/gll_grid.F90
	modified:   components/cam/src/physics/cam/phys_grid.F90
	modified:   components/cam/src/physics/cam/physics_types.F90