Fix calculation of CRM cloud fraction for rad
Fix the calculation of CRM cloud fraction that is intended for use in
the radiation module. Previously, this assumed that CF3D was
representing a proper 3D cloud fraction, i.e., that CF3D would be 0 if
there was no cloud, 1 if the CRM column was cloudy, and (possibly) 0 <
CF3D < 1 if the column was partially cloudy. However, this is only the
case if CLUBB is used, otherwise it is always CF3D = 1. This convention
appears to be used to make the code more general, but this means that we
have to add a check on the total cloud water to define cloud or no
cloud.