Allow setting iradsw/iradlw for SP
Allow setting iradsw and iradlw (frequency of shortwave and longwave
radiation) to something other than every GCM timestep (the old
behavior) to speed up the model when we need to run very small timesteps
for high resolution simulations, but cannot afford to run radiation that
frequently. Note that increasing the radiation timestep will decrease
the coupling between clouds and radiation, and at large radiation
timesteps the cloud radiative feedback will in all likelihood be
negatively affected. A new history field output is added for the
radiative heating applied to the CRM. This is collected at every GCM
timestep, even if it is not updated at that timestep, so that we can
confirm that we are getting the proper behavior (i.e., that CRM_QRAD is
not being zeroed out between radiation calls). Other CRM-resolved
radiative quantities are also fixed here to allow for output with the
reduced radiative grid (groups of columns).