Remove resetting of cld in crm_physics
Previously the crm_rad%cld variable was reset to zero based on a
threshold of total cloud water in crm_physics, with a note that this was
only used in the modal_water_uptake routine. It seems like this was done
to try to create a sort of effective cloud fraction of 0 or 1 for each
CRM column, but now we are intelligently computing a time-averaged cloud
fraction in crm_module, so this should be unnecessary, and setting this
within crm_module seems more appropriate. Leaving commented out for now,
but we can probably remove this altogether. However, this WILL cause
differences with master now.