Let modal_aero_wateruptake handle overcast assumption
Let modal_aero_wateruptake handle its own assumption about
clear/overcast, rather than using the crm_rad%cld definition. This
restores BFB behavior when SP_MCICA_RAD is unset. Our changes to how
crm_rad%cld was set previously broke BFB, because modal_aero_wateruptake
was using this variable in a somewhat strange way. This commit moves
handling of the clear/overcast assumption specific to
modal_aero_wateruptake to the routine it is used in, which seems to make
more sense. This also starts a small clean-up of the logic in that
routine related to SP.