Updated summit early science run scripts
Originally we planned for everything to go smoothly and we would
only need one run script for posterity / provenance (lol).
This updates the run scripts to reflect reality, and suggests a
new naming convention that corresponse to our case names.
Early science case names are appended with a datestring that should
match the datestring in the run script used for the original
submission.

 * early_science.sh has been renamed early_science.template.sh
 * early_science.20190222.sh has been added
 * early_science_branch.20190327.sh has been added for the branch
   run we will launch after fixing a restart problem from 20190222