Remove redundant add_hist_coord calls in phys_register
An earlier commit that moved the `add_hist_coord` calls for
CRM-level coordinates from `cam_diagnostics` to `crm_physics_register`
was duplicated by a later commit that moved these calls to
`phys_register`. The latter is removed by this commit, as it should
do the same thing as the former, and the former has already
been explicitly tested to validate it performs as intended.