Add 'vmag_gust' to pbuf in phys_register
Previously vmag_gust was added to pbuf in clubb_register_cam.
However, it is expected even when clubb is not in use. This is a
quick fix to ensure it is registered for our case, but this
should also be addressed upstream, since it may similarly
fail when moist_physics is not used.