Initial support for RRTMGP in SP
Add very minimal support for RRTMGP in SP configurations. This just adds
the CRM heating fields to the history buffer, and populates the CRM
heating with the net GCM heating, calculated from the GCM fields. That
is, homogeneous radiative heating is applied to the CRM, with that
heating calculated from the GCM domain average cloud properties, not the
CRM clould properties. This gets the code to compile and run, but is not
really appropriate yet.