Use CRM clouds in LW flux calcs
Initial support for using CRM cloud properties in the LW radiative flux
calculations. This implementation is a bit of a mess at the moment, but
the strategy is to pack all the CRM data into the ncol dimension of the
input arrays into the RRTMGP flux routines, and then to unpack/average
the data after calculating fluxes. This commit still keeps homogeneous
heating as output (it gets overwritten in radiation tend with domain
averages); that will be fixed in the next commit.