Simplify setting gas concentrations and pack
Refactor handling of gas concentrations to provide instead a routine
that sets a simple array that we can more easily manipulate at the
driver level. This allows us to pack the CRM data into a single
dimension more clearly. Not sure if this is the best approach yet, but
it does make it easier to copy the data into the bigger array because
the ty_gas_concs object that we returned before does not give direct
access into the underlying array without accessing via getter routines.