Fix a lot of bugs to make consistent with rrtmg
Fix a log of bugs in MMF configuration to bring answers into agreement
with both the RRTMG and the RRTMGP-loop implementation. TOA net fluxes
are not in good agreement between these three different radiation
implementations. This imports the Ebert-Curry and Slingo cloud optics
treatments from the add-sp-rrtmg-loop branch, so a lot of lines of code
were touched in this commit.