Do not collapse 4D outputs
A long time ago, it seems that CAM history did not support variables
with dimensions other than columns and height (i.e., only 2D/3D). Thus,
there was a bit of code in the COSP simulator interface that was needed
to collapse the 4D outputs before sending to CAM history. This is no
longer the case, so we can simplify the interface quite a bit and avoid
some unnecessary data movement.