Remove useless effective radius resetting
Previously, there was this check on the effective radius to see if it
was equal to the COSP fillvalue and would reset the converted effective
radius to fillvalue if true. However, the effective radius input from
the atmosphere model would NEVER be equal to COSP fillvalue, because the
atmosphere model does not even know about the COSP fillvalue, and
effective radius is never set to fillvalue anyway. So this block of code
was completely unnecessary.