Remove cosp_runall option
Remove the cosp_runall option. This was misleading and unnecessary. It
was used to set all simulators to on, but then also artifically forced
the sunlit flag to 1. This is misleading behavior we do not want to
support, and added unnecessary complication to the code and namelist
options. Note that this option was NOT needed for outputting the COSP
inputs for an offline driver, despite the comments in the code that said
that it was.