Fix handling of cosp_nradsteps namelist variable
Fix handling of cosp_nradsteps namelist variable so that we can easily
change this. The old behavior was to overwrite the value of
cosp_nradsteps if no simulator flags were present, because cosp_amwg
would be set which would then reset cosp_nradsteps to 3. So, for
example, if one were to build with `-docosp` but not explicitly set any
of the simulator logicals to `.true.`, then the code would set
`cosp_amwg = .true.` and overwrite all the namelist variables manually.
This commit simply sets a sane default for `cosp_nradsteps` (3) and then
allows overwriting if present on the namelist.