Add subcolumn outputs for MISR
Add outputs for subcolumn MISR-diagnosed cloud top heights and optical
depths. Requires working with a fork of the COSP2 repo to add these
fields to the COSP output types, but this is a worthwhile addition that
is useful for verifying the MISR simulator implementation, and also
gives access to the CRM-resolved MISR-simulated cloud top heights.