Only set nswbands and nlwbands in radconstants
Only set nswbands and nlwbands in radconstants (as parameters). Previously,
we had tried to set these dynamically in radiation by reading the values
from the kdist objects, which are populate from data read from netcdf files
at initialization, but some routines seemed to need these dimension sizes
to be compile-time constants. So, unfortunately, we revert to making the
values of nswbands and nlwbands hard-coded. Fixes #110