Add new stress bypass and surface scheme options
SP_STRESS_BYPASS_1: bypass application of stress in tphysac and add to
phys state BEFORE calling CRM. This makes the stress be part of the CRM
large-scale forcing.

SP_STRESS_BYPASS_2: bypass application of stress in tphysac and add to
 phys state AFTER calling CRM. This makes the stress NOT be part of the
CRM large-scale forcing.

SP_REMOVE_STRESS_FORCING: An alternative to the stress bypass options
due to results showing that the stress bypass leads to increases in GCM
wind. Instead of changing how the stress is applied, this option simply
subtracts the stress from the CRM large-scale momentum forcing so that
the SP_CRM_STRESS_SCHEME can add it to the CRM state in a heterogeneous
way.

SP_CRM_STRESS_SCHEME_HOM: Homogenously applies GCM surface stress to the
CRM wind. Does not take the CRM wind direction into account. Should only
be used when the stress is not included in the CRM large-scale forcing
as this would be double counting.

SP_CRM_STRESS_SCHEME_HET: Applies surface stress heterogenously across
CRM columns consistent with the direction of CRM wind and then applies
a correction to ensure the domain mean surface stress is identical
to the GCM stress.

	modified:   components/cam/src/physics/cam/crm_physics.F90
	modified:   components/cam/src/physics/cam/diffusion_solver.F90
	modified:   components/cam/src/physics/cam/physpkg.F90
	modified:   components/cam/src/physics/cam/restart_physics.F90
	modified:   components/cam/src/physics/cam/vertical_diffusion.F90
	modified:   components/cam/src/physics/crm/crm_module.F90
	modified:   components/cam/src/physics/crm/crmsurface.F90