Always use CRM-diagnosed cloud fraction
Always use the CRM-diagnosed cloud fraction rather than the (arbitrary)
column condensed water threshold to diagnose cloud. This does two
things. First and most importantly, it pulls a *partial* cloud fraction
from the CRM integration (currently the time average of a binary clear
or cloudy distinction at each CRM time step), which then enables
subcolumn sampling within each CRM column for the MCICA method. Note
that this should not add any additional cost, since this was already
being done, just with a trivial partial cloud fraction that was always 0
or 1. Second, it removes any implicitly assumed space or time scale for
the threshold against which to compare condensed water amount to
determine whether or not the entire column should be "clear" or
"cloudy", instead relying on the CRM to make this diagnosis.