Move SAM and ECPP code to subdirectory of crm
Move SAM and ECPP code to subdirectory of
components/cam/src/physics/crm, which will serve as the root location
for isolating superparameterization-specific code and create a separate
"crm" physics package (alongside the "cam" physics package). The code
that was in the crm directory is moved to a new "sam" subdirectory to
allow for the addition of other potential "crm" modules. `configure` is
also modified to enable building the code with this new organization.