Isolate CRM-specific code
Move routines that have been modified for MMF to the crm directory
and checkout versions from upstream E3SM in their place. This approach
isolates much of the MMF-specific code, and removes dependencies on CRM
dimensions and fields that have no meaning when building without MMF.
This also helps to more cleanly separate the CRM physics into a separate
physics "package", to be used as an alternative to the CAM/EAM physics
package.