Move consistency check on optics for rrtmg
Move consistency check on optics scheme for sam1mom/RRTMG out of
rad_constituents and into `radiation_init`. This allows us to use the
default `rad_constituents.F90` without modification for MMF. Handling of
the setting of `oldcldoptics` for sam1mom/RRTMG is moved to
`build-namelist`, and a consistency check is added to `radiation_init`
to make sure this is set correctly.