Remove oldcldoptics flag
Remove the defunct and misleading oldcldoptics flag. This flag had
inconsistent behavior between the shortwave and longwave optics. In the
shortwave, the flag would force the Ebert-Curry ice and the Slingo
liquid optics schemes to be used, but in the longwave the Ebert-Curry
and Slingo schemes would only be called to output diagnostics and then
would be overwritten by the Mitchell and Conley schemes. Thus, since we
were relying on the oldcldoptics flag for sam1mom MMF configurations, we
were accidentally relying on particle size information from the MG2
microphysics parameterization for sam1mom runs. Here, we replace the
use of the oldcldoptics flag by explicitly setting the optics scheme for
sam1mom and m2005 in `build-namelist`. Note that this will be non-BFB
for FSP1V1 configurations, but should be BFB for FSP2V1 and all other
configurations.