Add RRTMGP OpenACC kernels to build targets
Add paths for the OpenACC kernels to build targets so that we can build
the versions of these files that include the OpenACC directives for the
GPU. Note that these contain some files that are duplicated in the
regular kernels paths, but since they are added first, if those files
exist we use the OpenACC versions. This will not harm CPU builds, since
the code will compile and run in both places, it is just a little more
messy with the OpenACC directives scattered around and some workarounds
for compiler bugs or GPU considerations. Eventually these will hopefully
all merge to the main kernels directories, in which case we could remove
the extra paths, but there is no harm including them here.