Add OpenACC compiler flags for RRTMGP source
Add OpenACC compiler flags when building RRTMGP source with the summit
pgigpu compiler. We need to add a separate CMAKE variable to maintain
the list of RRTMGP files that need these flags, because the OpenACC
flags we are building SAM with caused problems for me. Specifically, it
looked like the `-Mcuda` flag was causing build issues. Simplifying the
compiler flags got around this problem, so we are using this approach
for now but should probably revisit to figure out exactly what was
causing the problems.