Use explicit optics types for rte_sw and rte_lw
Use explicit optics types for the rte_sw and rte_lw wrappers rather than
the polymorphic classes that the mo_rrtmgp_clr_all_sky routines used.
The abstract classes (ty_optical_props_arry) use seemed to be causing
problems with OpenACC, and replacing these with the explicit classes we
want to use (ty_optical_props_1scl for the longwave and
ty_optical_props_2str for the shortwave) fixed the problem, and the code
now builds and runs with OpenACC kernels on the GPU, and appears to give
the correct answers to within expected differences.