Move daytime compression to GPU
Move the compression of whole chunk arrays to daytime only arrays for
shortwave inputs to GPU by refactoring the data copying into kernels and
adding OpenACC directives. For the optical_props class instances, we
need to add a wrapper function that works on normal arrays, and pass the
individual arrays so that the compiler gets the pointer dereferencing
right on the GPU. Failing to do this resulted in Cuda runtime errors.